+++ src/frontend/web_application/src/scenes/ForgotPassword/presenter.jsx
@@ -44,9 +44,9 @@ class ForgotPassword extends Component {
     }
 
     const localizedErrors = {
-      [ERROR_IDENTIFIANTS_MISMATCH]: i18n._('passwords.form.error.identifiants_mismatch'),
-      [ERROR_USER_NOT_FOUND]: i18n._('passwords.form.error.user_not_found'),
-      [STATUS_INVALID_FORM]: i18n._('passwords.form.error.empty'),
+      [ERROR_IDENTIFIANTS_MISMATCH]: i18n._('passwords.form.error.identifiants_mismatch', { defaults: 'Identifiers don\'t match.' }),
+      [ERROR_USER_NOT_FOUND]: i18n._('passwords.form.error.user_not_found', { defaults: 'User not found.' }),
+      [STATUS_INVALID_FORM]: i18n._('passwords.form.error.empty', { defaults: 'At least one field is required.' }),
     };
 
     const errors = {