+++ src/frontend/web_application/src/scenes/Signup/form-validator.js
@@ -12,18 +12,18 @@ const ERR_INVALID_RECOVERY_EMAIL = 'ERR_INVALID_RECOVERY_EMAIL';
 const ERR_REQUIRED_RECOVERY_EMAIL = 'ERR_REQUIRED_RECOVERY_EMAIL';
 
 export const getLocalizedErrors = i18n => ({
-  [ERR_DOTS]: i18n._('signup.feedback.username_starting_ending_dot'),
-  [ERR_MIN_MAX]: i18n._('signup.feedback.username_length'),
-  [ERR_DOUBLE_DOTS]: i18n._('signup.feedback.username_double_dots'),
-  [ERR_REQUIRED_PRIVACY]: i18n._('signup.feedback.required_privacy'),
-  [ERR_REQUIRED_TOS]: i18n._('signup.feedback.required_tos'),
-  [ERR_INVALID_GLOBAL]: i18n._('signup.feedback.invalid'),
-  [ERR_REQUIRED_USERNAME]: i18n._('signup.feedback.required_username'),
-  [ERR_INVALID_CHARACTER]: i18n._('signup.feedback.username_invalid_characters', { 0: { chars: '"@`:;<>[]\\' } }),
-  [ERR_REQUIRED_PASSWORD]: i18n._('signup.feedback.required_password'),
-  [ERR_UNAVAILABLE_USERNAME]: i18n._('signup.feedback.unavailable_username'),
-  [ERR_INVALID_RECOVERY_EMAIL]: i18n._('signup.feedback.invalid_recovery_email'),
-  [ERR_REQUIRED_RECOVERY_EMAIL]: i18n._('signup.feedback.required_recovery_email'),
+  [ERR_DOTS]: i18n._('signup.feedback.username_starting_ending_dot', { defaults: 'The username cannot start or end with a dot (.)' }),
+  [ERR_MIN_MAX]: i18n._('signup.feedback.username_length', { defaults: 'The length of the username must be be between 3 and 42' }),
+  [ERR_DOUBLE_DOTS]: i18n._('signup.feedback.username_double_dots', { defaults: 'The username cannot contain two dots (.) next to the other' }),
+  [ERR_REQUIRED_PRIVACY]: i18n._('signup.feedback.required_privacy', { defaults: 'We need your privacy policy agreement' }),
+  [ERR_REQUIRED_TOS]: i18n._('signup.feedback.required_tos', { defaults: 'We need your terms and conditions agreement' }),
+  [ERR_INVALID_GLOBAL]: i18n._('signup.feedback.invalid', { defaults: 'Credentials are invalid' }),
+  [ERR_REQUIRED_USERNAME]: i18n._('signup.feedback.required_username', { defaults: 'A username is required' }),
+  [ERR_INVALID_CHARACTER]: i18n._('signup.feedback.username_invalid_characters', { values: { 0: '"@`:;<>[]\\' } }, { defaults: 'The username cannot contain some special characters like {0} and space' }),
+  [ERR_REQUIRED_PASSWORD]: i18n._('signup.feedback.required_password', { defaults: 'A password is required' }),
+  [ERR_UNAVAILABLE_USERNAME]: i18n._('signup.feedback.unavailable_username', { defaults: 'We are sorry, this username is not available' }),
+  [ERR_INVALID_RECOVERY_EMAIL]: i18n._('signup.feedback.invalid_recovery_email', { defaults: 'The email should be valid' }),
+  [ERR_REQUIRED_RECOVERY_EMAIL]: i18n._('signup.feedback.required_recovery_email', { defaults: 'A backup email is required' }),
 });
 
 const descriptor = {