+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - API for tags management : user's tags, tags on messages and contacts
+- add a `internal` tag to messages sent from the same instance than receiver
 
 ### Fixed